refactor : output 출력 부분 일부 변경
