+++ src/main/java/ladder/controller/LadderController.java
@@ -1,17 +1,17 @@
 package ladder.controller;
 
-import ladder.domain.ladder.Ladder;
+import ladder.domain.Ladder;
 import ladder.domain.user.Users;
 import ladder.view.LadderInput;
 import ladder.view.LadderOutput;
 
 public class LadderController {
 
-    public void start(){
+    public void start() {
         Users users = new Users(LadderInput.userAdd());
         int height = LadderInput.ladderHeight();
+        Ladder ladder = Ladder.of(height, users.countOfUser());
 
-        Ladder ladder = Ladder.of(height, users);
-        LadderOutput.drawLadder(ladder.getLadder());
+        LadderOutput.drawLadderGame(users.showUsers(), ladder.getLadder());
     }
 }
